Actually fix the problem with short timeouts
The buffer size for reading events was too small, leaving some
events still in the file descriptor after the read was complete.
This compounded, resulting in multiple seconds of old events being
backed up, resulting in rapidly flashing the backlight.  The
solution is to increase the buffer size.  Additionally, the input
thread sleep time is reduced to 500ms (the same value as
keylightd), which is necessary to handle the shortest timeouts.