Store the fade duration as nanoseconds instead of microseconds
Now that pthread_cond_timedwait() is used for the fade steps, it
no longer makes sense to store the fade duration as microseconds
and then need to convert it to nanoseconds every time the next fade
step time is calculated.  Instead, store it as nanoseconds and
convert it only once.