Swap input and timer threads again
Separating the fader and timer loops fixed the startup race that
previously required that input not be handled on the main thread.
Therefore, it is a good idea to switch back to handling input on
the main thread as this means that the shutdown signals can also
be handled on the main thread, which reduces complexity and allows
both created threads to be cancelled, which further reduces
complexity and increases efficiency by eliminating the shutdown
flag check.