Refactor fade initiation code
Several things happen that are common between starting to fade in
and out, so combine them into a function.  Additionally, reset
next_fade_step_time each time this happens to ensure that if one
fade is requested before the previous one completes, it begins
immediately instead of waiting for the previous interval to elapse.