Ignore EV_LED events
These are generated when scroll lock/caps lock/etc. gets turned on
or off with any keyboard.  These shouldn't cause the backlight to
turn on.  In order to make this possible, it is also necessary to
look at the second-to-last event in each received batch as the last
will always be an EV_SYN.  An additional check is added to ensure
that a non-conformant single event with no EV_SYN will not read off
the beginning of the array and cause undefined behavior or
crashing.