Improve EV_LED check
Since events are batched, it is possible that there could be events
we want to see followed by an EV_LED event in a single batch.  To
handle this, step backward through the read events until one that
is not an EV_SYN or EV_LED is found.  When one is found, break
immediately because we still only care about the latest event.  It
still makes sense to start at the second-to-last event because the
last event is always an EV_SYN.