Use pthread_cond_{timed}wait return value in fader loop
This improves efficiency by not having to compare timespecs on
every cycle.  It additionally makes the fade truly constant-time by
setting next_fade_step_time based on its previous value instead of
the current clock time.

This is not done for the timer loop because off_time may be changed
between when pthread_cond_timedwait was called and when it
returned, so off_time must always be compared to current_time to
ensure the the backlight should actually be turned off.  The other
alternative, waking up the timer thread on every input so that it
sees the new off_time, would have more overhead.