Avoid division by 0 in fade_init()
It is very unlikely, but if the timer thread turns off the
backlight (initiating a fade out) and then the input thread turns
on the backlight (initiating a fade in) before the fade out has
actually begun, fade_init() would try to divide by 0, crashing the
program.  Instead, check for that condition and set the fade
interval to 0 if that is the case.