Localize as many of the fade control variables as possible
This results in a small but tangible efficiency improvement.  This
works by "assuming" that if pthread_cond_{timed}wait() didn't
return ETIMEDOUT, it must have been signaled, which means that a
new fade has started.  (The fade_init variable is there just to
account for possible spurious wakeups.)  The separate check to
avoid division by 0 is no longer necessary because the code where
that could happen is already wrapped in a conditional that would
prevent it.