Rework the fader loop yet again to increase efficiency
Instead of checking for the direction of the fade on every step,
just calculate it once during fade initialization.  This reduces
the number of conditional evaluations during each fade step but
also allows the fade completion check to be moved after the
brightness is set in each step, avoiding the extra "ghost step" at
the end of each fade that was previously necessary to detect
completion.