Fix issue where backlight might turn off with keyboard input only
Because we only check for input immediately before the backlight
would otherwise be turned off, it is fully expected that the input
buffer will overflow and we will get EV_SYN SYN_DROPPED events.
However, when this occurs, the normal pattern of each actual event
being followed by an EV_SYN SYN_REPORT (therefore resulting in the
last event in the buffer always being EV_SYN SYN_REPORT) isn't
always followed.  With the previous code, if the buffer overflowed
at exactly the wrong time, this could cause a read to return 2
events with the first being an EV_SYN SYN_DROPPED.  This would
cause the event to be ignored, resulting in the backlight being
turned off at the wrong time.  To ensure this doesn't happen,
start at the very end of the buffer and allow EV_SYN SYN_DROPPED
events to set the timer.